domain_tracker/requirements.txt
aiofiles==24.1.0
alembic==1.13.2
bcrypt==4.0.1
bs4==0.0.2
celery==5.5.1
cryptography==43.0.0
fastapi==0.112.2
httpx==0.27.2
jinja2==3.1.4
langchain==0.2.15
markdown==3.7
passlib==1.7.4
psycopg2-binary==2.9.9
pytest==8.3.2
pytest-asyncio==0.24.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
python-whois==0.9.5
redis==5.2.1
SQLAlchemy==2.0.32
uvicorn==0.30.6
gunicorn==21.2.0
email-validator==2.2.0
apscheduler==3.10.4
